safari-600.7-branch/LayoutTests/fast/regions/cssom/element-region-overset-state-negative-letter-spacing-expected.txt

Tests Element.webkitRegionOverset attribute

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS region_1.webkitRegionOverset is "fit"
PASS region_2.webkitRegionOverset is "empty"
PASS successfullyParsed is true

TEST COMPLETE
